Move ImmutableFileCollection to new package
